refactor: improve fsd
