+++ src/pages/404.tsx
@@ -1,4 +1,4 @@
-import {Custom404Page} from "@/widgets";
+import {Custom404Page} from "@/widgets/custom-404-page";
 
 export default function Custom404() {
     return (
