+++ src/pages/_app.tsx
@@ -1,0 +1,5 @@
+import type { AppProps } from "next/app";
+
+export default function App({ Component, pageProps }: AppProps) {
+    return <Component {...pageProps} />;
+}
