+++ src/pages/api/logs.ts
@@ -10,7 +10,6 @@ export default function handler(req: NextApiRequest, res: NextApiResponse) {
         })
         res.status(200).json({message: "Log added"})
     } else if (req.method === 'GET') {
-        console.log(`\n\n\n\n\n\n\n ${getLogs().toString()}\n ${getLogs().length.toString()}\n\n\n\n\n`)
         res.status(200).json(getLogs());
     } else {
         res.status(405).end(`Method ${req.method} not allowed`);
