+++ src/widgets/book-info/book.module.css
