+++ src/widgets/book-list/ui/book-item/book-item.module.css
@@ -1,13 +1,3 @@
-.cards {
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 1rem;
-    padding: 1rem;
-    margin: 0;
-    background-color: rgba(146, 185, 142, 0.56);
-}
-
 .book {
     width: 100%;
     max-width: 15.1rem;
