+++ src/widgets/book-list/ui/index.ts
@@ -1,0 +1,1 @@
+export {BookList} from "./book-list"
