+++ src/widgets/custom-404-page/404.module.css
@@ -1,7 +1,7 @@
 .container {
     min-height: 100vh;
     width: 100%;
-    background-image: url('../../../../public/girl.png');
+    background-image: url('../../../public/girl.png');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
