+++ src/widgets/custom-404-page/index.tsx
@@ -1,4 +1,4 @@
-import styles from "@/widgets/сustom404Page/ui/404.module.css";
+import styles from "./404.module.css";
 
 export const Custom404Page = () => {
     return (
@@ -6,7 +6,7 @@ export const Custom404Page = () => {
             <div className={styles.content}>
                 <h1>404</h1>
                 <p>Sorry, page not found</p>
-                <a href="/">Back to main page</a>
+                <a href="/public">Back to main page</a>
             </div>
         </div>
     )
