+++ src/widgets/footer/footer.module.css
