--- src/widgets/index.ts
@@ -1,5 +1,0 @@
-export {Header} from "./header/ui/header"
-export {BookList} from "./bookList/ui/bookList"
-export {BookInfo} from "./bookInfo/ui/BookInfo"
-export {Custom404Page} from "./сustom404Page/ui/сustom404Page"
-export {Footer} from "./footer/ui/footer"
